doc: add doc to routes on swagger

+++ PasswordManager/Controllers/PasswordController.cs
@@ -1,4 +1,3 @@
-using System.ComponentModel;
 using Microsoft.AspNetCore.Mvc;
 using PasswordManager.Dto.Password;
 using PasswordManager.Entities;
@@ -8,19 +7,29 @@ namespace PasswordManager.Controllers
 {
     [ApiController]
     [Route("[controller]")]
+    [Produces("application/json")]
+    [Consumes("application/json")]
     public class PasswordController(IPasswordDataService service) : ControllerBase
     {
         private readonly IPasswordDataService _passwordService = service;
 
-
+        /// <summary>
+        /// Adds a New Password/Account to the Database
+        /// </summary>
+        /// <param name="request"></param>
+        /// <returns></returns>
         [HttpPost]
         [ProducesResponseType(typeof(int),StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)] 
         public IActionResult AddNewPassword([FromBody] RequestAddPasswordDataDTO request){
             var id = _passwordService.AddPassword(request);
 
             return Created(string.Empty,id);
         }
-
+        /// <summary>
+        /// Get a list if all passwords stored in the database
+        /// </summary>
+        /// <returns></returns>
         [HttpGet]
         [ProducesResponseType(typeof(List<PasswordData>),StatusCodes.Status200OK)]
         public IActionResult GetAllPasswords(){
@@ -28,11 +37,20 @@ namespace PasswordManager.Controllers
             return Ok(list);
         }
 
+        /// <summary>
+        /// Get a specific encrypted password details with their id.
+        /// </summary>
+        /// <param name="id">id of the password</param>
+        /// <returns>A response with the PasswordData</returns>
         [HttpGet]
         [Route("{id}")]
         [ProducesResponseType(typeof(PasswordData),StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string),StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string),StatusCodes.Status400BadRequest)]
         public IActionResult GetEncryptedPassword([FromRoute] int id){
+            if( id <= 0){
+                return BadRequest("Invalid Id requested!");
+            }
             var result = _passwordService.GetEncryptedPasswordData(id);
             if(result is null){
                 return NotFound("Password not found!");
@@ -42,11 +60,20 @@ namespace PasswordManager.Controllers
 
 
 
+        /// <summary>
+        /// Get a specific decrypted password details with their id.
+        /// </summary>
+        /// <param name="id">id of the password</param>
+        /// <returns>A response with the DecryptedPasswordData</returns>
         [HttpGet]
         [Route("Decrypt/{id}")]
         [ProducesResponseType(typeof(DecryptedPasswordData),StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string),StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string),StatusCodes.Status400BadRequest)]
         public IActionResult GetDecryptedPassword([FromRoute] int id){
+            if( id <= 0){
+                return BadRequest("Invalid Id requested!");
+            }
             var result = _passwordService.GetDecryptedPasswordData(id);
             if(result is null){
                 return NotFound("Password not found!");
@@ -54,11 +81,21 @@ namespace PasswordManager.Controllers
             return Ok(result);
         }
 
+        /// <summary>
+        /// Updates the content of a password
+        /// </summary>
+        /// <param name="id">id of the password</param>
+        /// <param name="request"></param>
+        /// <returns>returns the Updated PasswordData</returns>
         [HttpPut]
         [Route("{id}")]
         [ProducesResponseType(typeof(PasswordData),StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string),StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string),StatusCodes.Status400BadRequest)]
         public IActionResult UpdatePassword([FromRoute] int id, [FromBody] RequestAddPasswordDataDTO request){
+            if( id <= 0){
+                return BadRequest("Invalid Id requested!");
+            }
             var result = _passwordService.UpdatePasswordData(id,request);
             if(result is null){
                 return NotFound("Password not found!");
@@ -66,11 +103,20 @@ namespace PasswordManager.Controllers
             return Ok(result);
         }
 
+        /// <summary>
+        /// Deletes a password from the database
+        /// </summary>
+        /// <param name="id">id of the PasswordData</param>
+        /// <returns>Returns a message if deleted successfully or not</returns>
         [HttpDelete]
         [Route("{id}")]
         [ProducesResponseType(typeof(string),StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string),StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(string),StatusCodes.Status400BadRequest)]
         public IActionResult DeletePassword([FromRoute]int id) {
+            if( id <= 0){
+                return BadRequest("Invalid Id requested!");
+            }
             var result = _passwordService.DeletePasswordDataEntry(id);
 
             if(!result){
